templates/agent/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
httpx>=0.27
pydantic>=2.7
PyYAML>=6.0
# Опционально: включайте нужную LLM-интеграцию сами.
# openai>=1.30
# anthropic>=0.28
